Refactor: ModalWrapper.component 컨벤션을 맞춘다
