+++ src/components/common/index.ts
@@ -1,3 +1,3 @@
 export { default as Input } from './Input/Input';
 export { default as Button } from './Button/Button.component';
-export { default as ModalWrapper } from './ModalWrapper/ModalWrapper';
+export { default as ModalWrapper } from './ModalWrapper/ModalWrapper.component';
